+++ tests/Predis/Connection/MasterSlaveReplicationTest.php
@@ -576,7 +576,7 @@ class MasterSlaveReplicationTest extends PredisTestCase
         $connection = $this->getMock('Predis\Connection\SingleConnectionInterface');
 
         if ($parameters) {
-            $parameters = new ConnectionParameters($parameters);
+            $parameters = ConnectionParameters::create($parameters);
             $hash = "{$parameters->host}:{$parameters->port}";
 
             $connection->expects($this->any())